pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anki-enhance"
version = "0.1.0"
description = "Generate Anki flashcards from text using LLMs"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["anki", "flashcards", "llm", "language-learning", "spaced-repetition"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "google-genai>=1.0.0",
    "pyyaml>=6.0.0",
    "pypdf>=3.0.0",
    "genanki>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
anki-enhance = "anki_enhance.main:main"

[project.urls]
Homepage = "https://github.com/LyKex/anki-enhance"
Repository = "https://github.com/LyKex/anki-enhance.git"
Issues = "https://github.com/LyKex/anki-enhance/issues"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "ipython>=8.38.0",
]
